+++ src/entities/Review/ui/ReviewCard.tsx
@@ -1,21 +1,25 @@
-import styles from "./productCard.module.scss"
-import FullStar from "../vector/FullStar.tsx"
-import HalfStar from "../vector/HalfStar.tsx"
+import styles from "./reviewCard.module.scss"
+import FullStar from "../../../shared/ui/StarIcon/FullStar.tsx"
+import HalfStar from "../../../shared/ui/StarIcon/HalfStar.tsx"
 import React from "react";
 
-interface ProductCardProps {
-    title: string,
+interface ReviewCardProps {
+    username: string,
+    description: string,
     rating: number,
-    price: number
+    month: string,
+    day: number,
+    year: number
 }
 
-const ReviewCard:React.FC<ProductCardProps> = ({
-    title,
+const ReviewCard:React.FC<ReviewCardProps> = ({
+    username,
+    description,
     rating,
-    price,
-    img
-
-                                              }) => {
+    month,
+    day,
+    year
+}) => {
     const roundedRating = Math.floor(rating * 2) / 2;
     const renderStars = (rating: number) => {
         const stars = [];
@@ -40,15 +44,19 @@ const ReviewCard:React.FC<ProductCardProps> = ({
                     {rating}
                 </span>
             </div>
+            {/*Need to add logic to 3 dots*/}
             <div>
                 •••
             </div>
             <h2 className={styles.username}>
-                {title}
+                {username}
             </h2>
-            <h2>
-                {price}
-            </h2>
+            <p className={styles.review}>
+                {description}
+            </p>
+            <span className={styles.date}>
+                Posted on {month} {day}, {year}
+            </span>
         </article>
     )
 }
