+++ src/entities/Review/ui/reviewCard.module.scss
