+++ src/main.tsx
@@ -3,7 +3,7 @@ import { createRoot } from 'react-dom/client'
 import './index.css'
 import App from './App.tsx'
 import {Provider} from "react-redux";
-import {store} from "./store/store.ts"
+import {store} from "./app/providers/store/store.ts"
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
