+++ src/pages/HomePage/ui/HomePage.tsx
@@ -1,0 +1,18 @@
+import {FC} from 'react';
+import Header from "../../../widgets/Header/ui/Header.tsx"
+
+const HomePage: FC = () => {
+
+
+    return (
+
+        <>
+            <Header/>
+            <main>
+
+            </main>
+        </>
+    )
+}
+
+export default HomePage;
