+++ src/shared/ui/StarIcon/FullStar.tsx
