+++ src/shared/ui/StarIcon/HalfStar.tsx
