+++ src/widgets/Header/ui/Header.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { NavLink } from "react-router-dom";
 import "./header.module..scss"
-import Input from "../placeholder/Input.tsx";
+import Input from "../../../shared/ui/Input/Input.tsx";
 
 const Header: React.FC = () => {
     return (
